move fns in util.py to utils dir
